remove mrjars
