+++ modules/container/src/main/java/org/mule/runtime/container/internal/ContainerModulesDiscovererProvider.java
@@ -6,23 +6,32 @@
  */
 package org.mule.runtime.container.internal;
 
+import static org.mule.runtime.api.util.MuleSystemProperties.classloaderContainerJpmsModuleLayer;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
  * Has a factory method for obtaining a {@link ModuleDiscoverer}.
  * <p>
- * This is to be overridden when running in a JVM that supports JPMS (9+).
+ * This overrides the default class when running in a JVM that supports JPMS (9+), but has a fallback mechanism to provide the
+ * previous implementation.
  * 
  * @since 4.6
  */
 public class ContainerModulesDiscovererProvider {
 
+
   private static final Logger LOGGER = LoggerFactory.getLogger(ContainerModulesDiscovererProvider.class);
 
   public static ModuleDiscoverer containerModulesDiscoverer() {
-    LOGGER.debug("Default 'ContainerModulesDiscovererProvider' implementation, using 'ClasspathModuleDiscoverer'...");
-    return new ClasspathModuleDiscoverer();
+    if (classloaderContainerJpmsModuleLayer()) {
+      LOGGER.debug("MRJAR 'ContainerModulesDiscovererProvider' implementation, using 'JpmsModuleLayerModuleDiscoverer'...");
+      return new JpmsModuleLayerModuleDiscoverer(new ClasspathModuleDiscoverer());
+    } else {
+      LOGGER.debug("MRJAR 'ContainerModulesDiscovererProvider' implementation, using 'ClasspathModuleDiscoverer'...");
+      return new ClasspathModuleDiscoverer();
+    }
   }
 
 }
