+++ modules/container/src/main/java/org/mule/runtime/container/internal/JpmsModuleLayerModuleDiscoverer.java
@@ -92,7 +92,6 @@ public class JpmsModuleLayerModuleDiscoverer implements ModuleDiscoverer {
         .collect(toSet());
   }
 
-
   public class JpmsMuleContainerModule implements MuleContainerModule {
 
     private final Module jpmsModule;
