+++ modules/jar-handling-utils/pom.xml
@@ -12,97 +12,12 @@
     <artifactId>mule-jar-handling-utils</artifactId>
     <packaging>jar</packaging>
     <name>Mule JAR handling utils</name>
-    <description>
-        This is a multi release module that allows handling of JDK internal JAR classes needed for Java 8 runtime and a cleaner way in 11+.
-    </description>
+    <description>This is a module that allows handling of JDK internal JAR classes.    </description>
 
     <properties>
         <javaModuleName>org.mule.runtime.jar.handling.utils</javaModuleName>
         <formatterConfigPath>../../formatter.xml</formatterConfigPath>
-        <project.build.java11Sources>src/main/java11</project.build.java11Sources>
-        <project.build.java11SourceDirectory>${project.basedir}/${project.build.java11Sources}</project.build.java11SourceDirectory>
-        <project.build.java11TestSourceDirectory>${project.basedir}/src/test/java11</project.build.java11TestSourceDirectory>
     </properties>
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>net.revelc.code.formatter</groupId>
-                <artifactId>formatter-maven-plugin</artifactId>
-                <configuration>
-                    <directories>
-                        <directory>${project.build.sourceDirectory}</directory>
-                        <directory>${project.build.testSourceDirectory}</directory>
-                        <directory>${project.build.java11SourceDirectory}</directory>
-                        <directory>${project.build.java11TestSourceDirectory}</directory>
-                    </directories>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>default-compile</id>
-                        <goals>
-                            <goal>compile</goal>
-                        </goals>
-                        <configuration>
-                            <!-- In order to compile with full Java 8 compliance with access to jdk internal classes we need to clear the release 11 setting from parent pom -->
-                            <release combine.self="override"/>
-                            <source>${javaVersion}</source>
-                            <target>${javaVersion}</target>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>base-compile</id>
-                        <configuration>
-                            <release>11</release>
-                            <excludes combine.self="override"/>
-                            <compileSourceRoots>
-                                <compileSourceRoot>${project.build.java11SourceDirectory}</compileSourceRoot>
-                            </compileSourceRoots>
-                            <multiReleaseOutput>true</multiReleaseOutput>
-                        </configuration>
-                    </execution>
-                </executions>
-                <configuration>
-                    <encoding>ISO-8859-1</encoding>
-                    <proc>none</proc>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jar-plugin</artifactId>
-                <configuration>
-                    <archive>
-                        <manifestEntries>
-                            <Multi-Release>true</Multi-Release>
-                        </manifestEntries>
-                    </archive>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>build-helper-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <!-- Add the java 11 sources to the source artifact -->
-                        <id>attach-java11-sources</id>
-                        <!-- This has to run after compilation so it doesn't interfere with it (compler has its own MRJAR configuration) -->
-                        <phase>prepare-package</phase>
-                        <goals>
-                            <goal>add-source</goal>
-                        </goals>
-                        <configuration>
-                            <sources>
-                                <source>${project.build.java11Sources}</source>
-                            </sources>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
-    </build>
 
     <dependencies>
         <dependency>
