+++ modules/service/pom.xml
@@ -15,77 +15,7 @@
         <javaModuleName>org.mule.runtime.service</javaModuleName>
         <skipExportTests>false</skipExportTests>
         <formatterConfigPath>../../formatter.xml</formatterConfigPath>
-        <project.build.java11SourceDirectory>${project.basedir}/src/main/java11</project.build.java11SourceDirectory>
-        <project.build.java11TestSourceDirectory>${project.basedir}/src/test/java11</project.build.java11TestSourceDirectory>
     </properties>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>net.revelc.code.formatter</groupId>
-                <artifactId>formatter-maven-plugin</artifactId>
-                <configuration>
-                    <directories>
-                        <directory>${project.build.sourceDirectory}</directory>
-                        <directory>${project.build.testSourceDirectory}</directory>
-                        <directory>${project.build.java11SourceDirectory}</directory>
-                        <directory>${project.build.java11TestSourceDirectory}</directory>
-                    </directories>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>compile-java-11</id>
-                        <phase>compile</phase>
-                        <goals>
-                            <goal>compile</goal>
-                        </goals>
-                        <configuration>
-                            <release>11</release>
-                            <compileSourceRoots>
-                                <compileSourceRoot>${project.build.java11SourceDirectory}</compileSourceRoot>
-                            </compileSourceRoots>
-                            <multiReleaseOutput>true</multiReleaseOutput>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jar-plugin</artifactId>
-                <configuration>
-                    <archive>
-                        <manifestEntries>
-                            <Multi-Release>true</Multi-Release>
-                        </manifestEntries>
-                    </archive>
-                </configuration>
-            </plugin>
-	        <plugin>
-	            <groupId>org.codehaus.mojo</groupId>
-	            <artifactId>build-helper-maven-plugin</artifactId>
-	            <executions>
-	                <execution>
-                        <!-- Add the java 11 sources to the source artifact -->
-	                	<id>attach-java11-sources</id>
-	                	<!-- This has to run after compilation so it doesn't interfere with it (compler has its own MRJAR configuration) -->
-	                    <phase>prepare-package</phase>
-	                    <goals>
-	                        <goal>add-source</goal>
-	                    </goals>
-	                    <configuration>
-	                        <sources>
-	                            <source>src/main/java11</source>
-	                        </sources>
-	                    </configuration>
-	                </execution>
-	            </executions>
-	        </plugin>
-        </plugins>
-    </build>
 
     <dependencies>
         <dependency>
